Organized project files into structured folders
