+++ src/main.jsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
+import "./Styles/index.css";
 import App from './App.jsx'
 
 createRoot(document.getElementById('root')).render(
@@ -8,3 +8,7 @@ createRoot(document.getElementById('root')).render(
     <App />
   </StrictMode>,
 )
+
+git add .
+git commit -m "Added navbar and home page"
+git push
